Safari-611.1.10/LayoutTests/crypto/subtle/hmac-import-jwk-key-sha512-expected.txt

Test importing a JWK HMAC key with SHA-512

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 128
PASS key.algorithm.hash.name is 'SHA-512'
PASS key.usages is ['sign', 'verify']
PASS successfullyParsed is true

TEST COMPLETE
